requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
boto3
python-dotenv